test: get sudo permissions for sniffing packets
As a workaround, the script will check the EUID and show the sudo
prompt to elevate permissions for the packet sniffing. That works,
but the actual test does not work yet.

+++ tests/test.py
@@ -1,7 +1,16 @@
 import subprocess
+import sys
+
 from yaspin import yaspin
 import time
 from scapy.all import *
+
+def check_permissions():
+    return os.geteuid() == 0
+
+def sudo():
+    args = [sys.executable] + sys.argv
+    os.execlp('sudo', 'sudo', *args)
 
 def run():
     spawn = subprocess.Popen('../build/main', stdout=subprocess.PIPE, stderr=subprocess.PIPE)
@@ -13,17 +22,26 @@ def check_broadcast(packet):
             return True
         return False
 
-process = run()
+def main():
+    if not check_permissions():
+        print("This script, as a workaround, requires sudo privileges. Restarting...")
+        sudo()
+        return
 
-with yaspin(text="Checking for broadcast message...", color="yellow") as spinner:
-    start_time = time.time()
-    while time.time() - start_time < 30:
-        sniff = sniff(filter="udp and port 9090", timeout=1, count=1)
-        if sniff and check_broadcast(sniff[0]):
-            spinner.ok("OK")
-            break
-    else:
-        spinner.fail("FAIL")
+    process = run()
 
-process.terminate()
-process.wait()
+    with yaspin(text="Checking for broadcast message...", color="yellow") as spinner:
+        start_time = time.time()
+        while time.time() - start_time < 30:
+            sniffer = sniff(filter="udp and port 9090", timeout=1, count=1)
+            if sniffer and check_broadcast(sniff[0]):
+                spinner.ok("OK")
+                break
+        else:
+            spinner.fail("FAIL")
+
+    process.terminate()
+    process.wait()
+
+if __name__ == "__main__":
+    main()
